Had to call replay history from blueprint (by the way the only thing it does now is create a new struct array - will consider renaming) and implement the loop around replay action (still implemented in c++) with blueprint nodes. I may have to consider making both funcitons plainly BlueprintCallable and nothing more. Had to set all floats in the new struct to zero; that helped the Jump value to not get "stuck" at 1 after the first trigger.
Complete replication of movement seems to be working now. Need to run further tests for accuracy. Turns out I didn't need to delay by a specific amount when going through the array of structs that replicate the movements/actions. I do need to add a minute delay in the loop though which seems to force the loop to execute once every tick, otherwise if it's an immediate loop then all actions are executed in one entire frame, which is too fast to notice anything! I have to understand why it works this way in order to figure out a more robust solution.