First rudimentary working version of the TArray implementation:
Interaction Commands are stacked in an array. This enables swapping for example of various props when using the GrabInteraction. Could be used to combine actions with 2 props (maybe more?), for example lightin a brazier while holding a torch (in the future unlocking a door while holding a key? etc.)

Still some code cleanup necessary (deleting superfluous UE_LOGs or commented out blocks of older code etc.)
For now I have disabled the attachment of attachable props, to make the swapping of props implementation easier.
Also used a debugging printscreen with keyboard input "I" to print the stack of interaction commands (palced in ChronoCharacter's Blueprint, should be removed as soon as practical)